backend/requirements.txt
Flask>=3.1.2
flask-cors>=4.0.0
python-dotenv>=1.1.1
pdfminer.six>=20250506
python-docx>=1.2.0
requests>=2.32.5